Idea I had for a way to manage data
this is just a header file for a state_row class I worked on for like 5 min.
I was brainstorming for a bit tonight but scrapped most of what I wrote. I felt myself making to 
many structural choices without anyone's input. So here is one of the ideas I had in it's barest form.